refactor: 공통 응답 객체 위치 수정
